perf(davinci): route sfc dom templates through s2 sections
